fix conflict

+++ Control/PID/PID.h
@@ -5,21 +5,21 @@
 
 struct PID
 {
-	uint16_t kP;
-	uint16_t kI;
-	uint16_t kD;
-	uint16_t dt;
-	int32_t P;
-	int32_t I;
-	int32_t D;
-	int32_t set;
-	int32_t feedback;
-	int32_t last_error;
-	int32_t error;
-	int32_t derror;
+	float kP;
+	float kI;
+	float kD;
+	float dt;
+	float P;
+	float I;
+	float D;
+	float set;
+	float feedback;
+	float last_error;
+	float error;
+	float derror;
 	int32_t Imax;
 	int32_t outmax;
-	int32_t out;
+	float out;
 };
 
 
@@ -27,9 +27,9 @@ int PID_ROLL(int roll);
 int PID_YAW(int yaw);
 int PID_PITCH(int pitch);
 
-void PIDdataInit(struct PID* pidData,uint16_t kp,uint16_t ki,uint16_t kd,int32_t imax,int32_t outmax);
+void PIDdataInit(struct PID* pidData,float kp,float ki,float kd,int32_t imax,int32_t outmax);
 void PIDdataUpdate(struct PID* pidData);
-void Reset_set(struct PID* pidData, int32_t Newset);
+void Reset_set(struct PID* pidData, float Newset);
 void ResetI(struct PID* pidData);
 void GetP(struct PID* pidData);
 void GetI(struct PID* pidData);
